refactor: move task status update tracking logic to components
