+++ physical_ai_manager/src/components/InfoPanel.js
@@ -22,6 +22,7 @@ import toast from 'react-hot-toast';
 import { MdVisibility, MdVisibilityOff } from 'react-icons/md';
 import { useSelector, useDispatch } from 'react-redux';
 import { setTaskInfo } from '../features/tasks/taskSlice';
+import TaskPhase from '../constants/taskPhases';
 
 const taskInfos = [
   {
@@ -54,12 +55,18 @@ const taskInfos = [
   },
 ];
 
-const InfoPanel = ({ disabled = false }) => {
-  const info = useSelector((state) => state.tasks.taskInfo);
+const InfoPanel = () => {
   const dispatch = useDispatch();
+
+  const info = useSelector((state) => state.tasks.taskInfo);
+  const taskStatus = useSelector((state) => state.tasks.taskStatus);
+
+  const [isTaskStatusPaused, setIsTaskStatusPaused] = useState(false);
+  const [lastTaskStatusUpdate, setLastTaskStatusUpdate] = useState(Date.now());
 
   const [showPopup, setShowPopup] = useState(false);
   const [taskInfoList] = useState(taskInfos);
+  const disabled = taskStatus.phase !== TaskPhase.READY || !isTaskStatusPaused;
   const [isEditable, setIsEditable] = useState(!disabled);
 
   // User ID list for dropdown
@@ -168,6 +175,28 @@ const InfoPanel = ({ disabled = false }) => {
       handleUserIdSelect(userIdList[0]);
     }
   }, [userIdList, info.userId, handleUserIdSelect]);
+
+  // track task status update
+  useEffect(() => {
+    if (taskStatus) {
+      setLastTaskStatusUpdate(Date.now());
+      setIsTaskStatusPaused(false);
+    }
+  }, [taskStatus]);
+
+  // Check if task status updates are paused (considered paused if no updates for 1 second)
+  useEffect(() => {
+    const UPDATE_PAUSE_THRESHOLD = 1000;
+    const timer = setInterval(() => {
+      const timeSinceLastUpdate = Date.now() - lastTaskStatusUpdate;
+      const isPaused = timeSinceLastUpdate >= UPDATE_PAUSE_THRESHOLD;
+      if (isPaused !== isTaskStatusPaused) {
+        setIsTaskStatusPaused(isPaused);
+      }
+    }, 1000);
+
+    return () => clearInterval(timer);
+  }, [lastTaskStatusUpdate, isTaskStatusPaused]);
 
   const classLabel = clsx('text-sm', 'text-gray-600', 'w-28', 'flex-shrink-0', 'font-medium');
